+++ CutTheRope/game/CTRMover.cs
@@ -14,10 +14,10 @@ namespace CutTheRope.game
             if (p.characterAtIndex(0) == 'R')
             {
                 bool flag = p.characterAtIndex(1) == 'C';
-                int num = (int)RTD((double)p.substringFromIndex(2).intValue());
+                int num = (int)RTD(p.substringFromIndex(2).intValue());
                 num *= 3;
                 int num2 = num / 2;
-                float num3 = (float)(6.283185307179586 / (double)num2);
+                float num3 = (float)(6.283185307179586 / num2);
                 if (!flag)
                 {
                     num3 = 0f - num3;
@@ -25,8 +25,8 @@ namespace CutTheRope.game
                 float num4 = 0f;
                 for (int i = 0; i < num2; i++)
                 {
-                    float x = s.x + (float)num * cosf(num4);
-                    float y = s.y + (float)num * sinf(num4);
+                    float x = s.x + num * cosf(num4);
+                    float y = s.y + num * sinf(num4);
                     addPathPoint(vect(x, y));
                     num4 += num3;
                 }
