+++ CutTheRope/game/GameController.cs
@@ -54,25 +54,25 @@ namespace CutTheRope.game
             gameScene.gameSceneDelegate = this;
             gameView.addChildwithID(gameScene, 0);
             Button button = MenuController.createButtonWithImageQuad1Quad2IDDelegate(69, 0, 1, 6, this);
-            button.x = (float)-(float)canvas.xOffsetScaled;
+            button.x = -(float)canvas.xOffsetScaled;
             gameView.addChildwithID(button, 1);
             Button button2 = MenuController.createButtonWithImageQuad1Quad2IDDelegate(62, 0, 1, 1, this);
-            button2.x = (float)-(float)canvas.xOffsetScaled;
+            button2.x = -(float)canvas.xOffsetScaled;
             gameView.addChildwithID(button2, 2);
             Image image = Image.Image_createWithResIDQuad(66, 0);
             image.anchor = image.parentAnchor = 10;
             image.scaleX = image.scaleY = 1.25f;
-            image.rotationCenterY = (float)(-(float)image.height / 2);
+            image.rotationCenterY = -(float)image.height / 2;
             image.passTransformationsToChilds = false;
             mapNameLabel = new Text().initWithFont(Application.getFont(4));
             mapNameLabel.setName("mapNameLabel");
             CTRRootController cTRRootController = (CTRRootController)Application.sharedRootController();
             CTRPreferences.getScoreForPackLevel(cTRRootController.getPack(), cTRRootController.getLevel());
             mapNameLabel.anchor = mapNameLabel.parentAnchor = 12;
-            mapNameLabel.x = RTD(-10.0) - (float)canvas.xOffsetScaled + 256f;
+            mapNameLabel.x = RTD(-10.0) - canvas.xOffsetScaled + 256f;
             mapNameLabel.y = RTD(-5.0);
             image.addChild(mapNameLabel);
-            VBox vBox = new VBox().initWithOffsetAlignWidth(5.0, 2, (double)SCREEN_WIDTH);
+            VBox vBox = new VBox().initWithOffsetAlignWidth(5.0, 2, SCREEN_WIDTH);
             Button c = MenuController.createButtonWithTextIDDelegate(Application.getString(655397), 0, this);
             vBox.addChild(c);
             Button c2 = MenuController.createButtonWithTextIDDelegate(Application.getString(655398), 2, this);
@@ -85,11 +85,11 @@ namespace CutTheRope.game
             Vector offset = vectSub(Image.getQuadCenter(8, 0), Image.getQuadOffset(8, 12));
             ToggleButton toggleButton = MenuController.createAudioButtonWithQuadDelegateIDiconOffset(3, this, 10, vectZero);
             ToggleButton toggleButton2 = MenuController.createAudioButtonWithQuadDelegateIDiconOffset(2, this, 11, offset);
-            HBox hBox = new HBox().initWithOffsetAlignHeight(-10f, 16, (float)toggleButton.height);
+            HBox hBox = new HBox().initWithOffsetAlignHeight(-10f, 16, toggleButton.height);
             hBox.addChild(toggleButton2);
             hBox.addChild(toggleButton);
             vBox.addChild(hBox);
-            vBox.y = (SCREEN_HEIGHT - (float)vBox.height) / 2f;
+            vBox.y = (SCREEN_HEIGHT - vBox.height) / 2f;
             bool flag3 = Preferences._getBooleanForKey("SOUND_ON");
             bool flag2 = Preferences._getBooleanForKey("MUSIC_ON");
             if (!flag3)
@@ -324,7 +324,7 @@ namespace CutTheRope.game
             switch (n)
             {
                 case 0:
-                    ((GameScene)view.getChild(0)).dimTime = (float)tmpDimTime;
+                    ((GameScene)view.getChild(0)).dimTime = tmpDimTime;
                     tmpDimTime = 0;
                     setPaused(false);
                     CTRRootController.logEvent("IM_CONTINUE_PRESSED");
@@ -664,9 +664,9 @@ namespace CutTheRope.game
         public override void fullscreenToggled(bool isFullscreen)
         {
             View view = getView(0);
-            view.getChild(2).x = (float)-(float)canvas.xOffsetScaled;
-            view.getChild(1).x = (float)-(float)canvas.xOffsetScaled;
-            mapNameLabel.x = RTD(-10.0) - (float)canvas.xOffsetScaled + 256f;
+            view.getChild(2).x = -(float)canvas.xOffsetScaled;
+            view.getChild(1).x = -(float)canvas.xOffsetScaled;
+            mapNameLabel.x = RTD(-10.0) - canvas.xOffsetScaled + 256f;
             GameScene gameScene = (GameScene)view.getChild(0);
             gameScene?.fullscreenToggled(isFullscreen);
         }
